+++ src/python/datapreprocessor/datapreprocessor/program.py
@@ -1,6 +1,7 @@
 import argparse
 from csvreader import *
 from datanormaliser import *
+from csvcleanser import *
 
 # This script reads in one or many csv files containing information about minstry meetings.
 # It assumes that the files have the following four ordered columns:
@@ -31,9 +32,7 @@ def print_lines(lines):
 
 def get_csv_file_lines(filename):
     file_contents = read_file(filename)
-    file_contents = remove_empty_lines(file_contents)
-    file_contents = remove_empty_columns(file_contents)
-    file_contents = remove_boilerplate(file_contents)
+    file_contents = cleanse_csv_data(file_contents)
     file_contents = clean_minister_column(file_contents, filename)
     file_contents = clean_dates(file_contents)
     return file_contents
